cleanup test code
